refactor release + ad hoc prerelease
