--- .github/workflows/pr-deploy.yml
@@ -1,16 +1,0 @@
-name: PR deploy
-
-on:
-    workflow_dispatch:
-        inputs:
-            pr:
-                description: 'PR number'
-                required: true
-
-jobs:
-    print:
-        runs-on: ubuntu-latest
-
-        steps:
-        - run: |
-            echo "PR number is ${{github.event.inputs.pr}}"
